+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.18.2] - 2022-12-07
 ### Changed
+- Improved the POST notification process by validating the request, before retrieving the order
 - Improved the error message that is shown to the customer whenever the payment fails.
 - Checking for recurring tokens will only occur now when Magento Vault is enabled.